Shorten twirp handler abstract method param from request to req
Crystal 1.5 warns if a method is overridden with mismatching parameter names (https://github.com/crystal-lang/crystal/pull/11915). While quite sensible, this unfortunately means that this shard needs to have an opinion on what a good name for this param should be. Given that twirp handlers tend to access request parameters a lot, `request` felt rather verbose.