Use a connection pool
This commit uses the DB::Pool class for a thread-safe, configurable
connection pool. By default, it spins up a single connection on
instantiation and will close all but one connection on checking
connections back into the pool. This commit does not add configurability
to that connection pool, but it can be added in a followup commit.